requirements.txt
langchain-community
langchain
faiss-cpu
torch
transformers